tidy(mm): add default_settings to model config
